requirements.txt
requests == 2.27.1
pyyaml == 6.0
tinydb == 4.5.2
boto3 == 1.20.33
tabulate == 0.8.9
cachetools == 5.0.0

notebook == 6.4.8
matplotlib == 3.5.1
numpy == 1.22.2
